+++ src/protocol.h
@@ -146,7 +146,7 @@ typedef struct SyncEntry
   int sync_start_height;
 
   uint32_t last_sync_height;
-  int64_t last_sync_ts;
+  uint32_t last_sync_ts;
   uint8_t last_sync_tries;
 } sync_entry_t;
