+++ src/util.c
@@ -110,7 +110,7 @@ int make_hash(char *digest, unsigned char *string)
   return 0;
 }
 
-int64_t get_current_time(void)
+uint32_t get_current_time(void)
 {
-  return (int64_t)time(NULL);
+  return (uint32_t)time(NULL);
 }
